chore: FullscreenVideoContainer 파일 분리
